test: hashing names of test ref files to appease R CMD CHECK.
Essentially there is a limitation on file path length in tar and, since tarballs are used to create a built package artifact, this limitation is passed through to R CMD CHECK. Some of our (programmatically built) test ref files had names that were too long. We now do an MD5 hash of the name before saving the file, in order to avoid this issue.